Stable-Version-1.0.5
+Added link obliterator
+Fixed bugs and deprecation issues with Gatekeeper
+Message now tells user if incorrect

+++ AnthaBot.js
@@ -27,8 +27,16 @@ client.login(config.token)
 
 client.on('message', msg => { 
     const CurrentTime = new Date();
-    //Below is the prevention of Botception.
+    //Prevention of Botception.
     if (msg.author.bot) return
+    
+    //Link obliterator
+    if (msg.content.includes("https://discord.gg/")){
+        msg.delete()
+        msg.member.send('You are reminded that outside Discord server links are not permitted in the server.')
+        console.log(`${member.displayName} has had it's Discord link obliterated on ${CurrentTime}`)
+    }
+
     //User-initiated commands
     //PING
     if (msg.content.startsWith(config.prefix + "ping")) {
@@ -82,9 +90,6 @@ client.on('message', msg => {
         GuildID.fetchMember(msg.author.id).then (member => {
             if (member.roles.has(MemberRole.id)) return
             if(!member.roles.has(MemberRole.id)){
-                if (!msg.content.includes() == 'agree'){
-                    msg.author.send("Incorrect! Please Try Again! Remember, it's only ONE word, NOTHING ELSE. If you include other words, **I will not recognize it.**")
-                }
                 member.addRole(MemberRole)
                 msg.author.send(`Congratulations! You are now a member of the server! Enjoy your stay! My suffix is !Yo!, and your commands as a member are: "!Yo!Anthabot, <Any Yes/no question>", "!Yo!Ping" ,"!Yo!Hello!", "!Yo!How Are you?"`)
                 .then(() => {
@@ -94,22 +99,33 @@ client.on('message', msg => {
             }
         }
     )}
-
-
-    if (msg.content.toLowerCase() == 'agree' && msg.channel.id === '404305206743007254') {
+    if (!((msg.content.includes('agree')) || (msg.content.includes('Agree'))) && (msg.channel.type === 'dm')){
+        
+        GuildID.fetchMember(msg.author.id).then (member => {
+            if(member.roles.has(MemberRole.id)) return
+            if(!member.roles.has(MemberRole.id)){
+                msg.author.send("Incorrect! Please Try Again! Remember, it's only ONE word, NOTHING ELSE. If you include other words, **I will not recognize it.**")
+            }
+        }
+    )
+}
+        
+    if (msg.content.toLowerCase() == 'agree' && msg.channel.id === '404305206743007254') {     
         if (msg.member.roles.has(MemberRole.id)) return
             if(!msg.member.roles.has(MemberRole.id)){
-                if (!msg.content.includes() == 'agree'){
-                    msg.member.send("Incorrect! Please Try Again! Remember, it's only ONE word, NOTHING ELSE. If you include other words, **I will not recognize it.**")
-                    msg.delete()
-                }
-        }
-        msg.member.addRole(MemberRole)
-        msg.delete()
-        msg.member.send(`Congratulations! You are now a member of the server! Enjoy your stay! My suffix is !Yo!, and your commands as a member are: "!Yo!Anthabot, <Any Yes/no question>", "!Yo!Ping" ,"!Yo!Hello!", "!Yo!How Are you?"`)
-        .then(() => {
+                msg.delete()
+                msg.member.addRole(MemberRole)
+            msg.member.send(`Congratulations! You are now a member of the server! Enjoy your stay! My suffix is !Yo!, and your commands as a member are: "!Yo!Anthabot, <Any Yes/no question>", "!Yo!Ping" ,"!Yo!Hello!", "!Yo!How Are you?"`)
+            .then(() => {
             msg.guild.channels.get('404304757558345739').send(`${msg.member.displayName} has been verified and confirmed as a new member! Please welcome them to the server!`)
             console.log(`${msg.member.displayName} has been verified via Guild Channel on ${CurrentTime}`)
         })
+        }
+     }
+     if (!msg.content.includes('agree') && msg.channel.id === '404305206743007254'){
+        msg.delete()
+        msg.member.send("Incorrect! Please Try Again! Remember, it's only ONE word, NOTHING ELSE. If you include other words, **I will not recognize it.**")
+     }
     }
-})
+    
+)
